# api/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Olve.Trains.UI.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Olve.Trains.UI.Server"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/run-command": {
      "post": {
        "tags": [
          "Olve.Trains.UI.Server"
        ],
        "operationId": "RunCommand",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultProblem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/logs": {
      "get": {
        "tags": [
          "Olve.Trains.UI.Server"
        ],
        "operationId": "GetLogs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogMessage"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultProblem"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Exception": {
        "type": "object",
        "properties": {
          "targetSite": {
            "$ref": "#/components/schemas/MethodBase"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "innerException": {
            "$ref": "#/components/schemas/Exception"
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "hResult": {
            "type": "integer",
            "format": "int32"
          },
          "stackTrace": {
            "type": "string",
            "nullable": true
          }
        },
        "nullable": true
      },
      "IPath": { },
      "IPath2": {
        "nullable": true
      },
      "LogLevel": {
        "type": "integer"
      },
      "LogMessage": {
        "required": [
          "level",
          "message",
          "sourcePath",
          "sourceLine",
          "time",
          "tags"
        ],
        "type": "object",
        "properties": {
          "level": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "message": {
            "type": "string"
          },
          "sourcePath": {
            "$ref": "#/components/schemas/IPath2"
          },
          "sourceLine": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
      "MethodBase": {
        "nullable": true
      },
      "ProblemOriginInformation": {
        "type": "object",
        "properties": {
          "filePath": {
            "$ref": "#/components/schemas/IPath"
          },
          "lineNumber": {
            "type": "integer",
            "format": "int32"
          },
          "linkString": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ResultProblem": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "severity": {
            "type": "integer",
            "format": "int32"
          },
          "args": {
            "type": "array",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "$ref": "#/components/schemas/Exception"
          },
          "originInformation": {
            "$ref": "#/components/schemas/ProblemOriginInformation"
          }
        }
      },
      "RunCommandRequest": {
        "required": [
          "command"
        ],
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Olve.Trains.UI.Server"
    }
  ]
}